+++ test/app/musicplayer/model/PlaylistTest.java
@@ -1,16 +1,10 @@
 package app.musicplayer.model;
 
-import app.musicplayer.model.Playlist;
-import app.musicplayer.model.Song;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import org.junit.jupiter.api.Test;
-import org.mockito.Mockito;
 
-import java.lang.reflect.Constructor;
 import java.lang.reflect.InvocationTargetException;
-import java.lang.reflect.Method;
-import java.lang.reflect.Modifier;
 import java.time.Duration;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
