+++ test/app/musicplayer/model/SearchResultTest.java
@@ -1,11 +1,9 @@
 package app.musicplayer.model;
 
-import app.musicplayer.model.*;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.testfx.framework.junit5.ApplicationExtension;
 
-import java.lang.reflect.InvocationTargetException;
 import java.time.Duration;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
