+++ test/app/musicplayer/view/ControlSongMoment.java
@@ -1,4 +1,4 @@
-package app.musicplayer.model;
+package app.musicplayer.view;
 
 import app.musicplayer.MusicPlayer;
 import javafx.application.Platform;
@@ -14,12 +14,12 @@ import org.sikuli.script.Screen;
 import org.testfx.framework.junit5.ApplicationExtension;
 
 @ExtendWith(ApplicationExtension.class)
-public class ControlSongInfo {
+public class ControlSongMoment {
 
     @Test
-    public void SongInfo() throws InterruptedException, FindFailed {
+    public void SongMoment() throws InterruptedException, FindFailed {
         Screen screen = new Screen();
-        ImagePath.add("test\\resources\\ControlSongInfo");
+        ImagePath.add("test\\resources\\ControlSongBar");
         Thread thread = new Thread(new Runnable() {
             @Override
             public void run() {
@@ -38,8 +38,10 @@ public class ControlSongInfo {
             }
         });
         thread.start();
-        screen.click("songs.png");
-        Thread.sleep(3000);
-        assertNotNull(screen.exists("songinfo.png"));
+        screen.click("play.png");
+        Thread.sleep(5000);
+        screen.dragDrop("SongBar.png", "restart.png");
+        Thread.sleep(1000);
+        assertNotNull(screen.exists("end.png"));
     }
 }
