+++ test/app/musicplayer/view/ControlVolumeTest.java
@@ -1,4 +1,4 @@
-package app.musicplayer.model;
+package app.musicplayer.view;
 
 import app.musicplayer.MusicPlayer;
 import javafx.application.Platform;
